Check app module
Overlay not working. This branch (file-overlay) is so that I can
incrementally add back the code from the supposedly completed branch on
the overlay tutorial i.e. the one not working, and see if I can step
through adding all the changes in an order that keeps the ng serve from
going to a blank screen (which is what I have on the current
Overlay-Screenplay branch) and make any changes that I see necessary on
the Overlay-Screenplay branch.